configurable trace sampling
