+++ config/crds/istio_v1beta1_istio.yaml
@@ -132,6 +132,9 @@ spec:
                 replicaCount:
                   format: int32
                   type: integer
+                traceSampling:
+                  format: float
+                  type: number
               type: object
             proxy:
               description: Proxy configuration options
