+++ pkg/apis/istio/v1beta1/remoteistio_types.go
@@ -103,23 +103,3 @@ type RemoteIstioList struct {
 func init() {
 	SchemeBuilder.Register(&RemoteIstio{}, &RemoteIstioList{})
 }
-
-func SetRemoteIstioDefaults(remoteconfig *RemoteIstio) {
-	if remoteconfig.Spec.IncludeIPRanges == "" {
-		remoteconfig.Spec.IncludeIPRanges = defaultIncludeIPRanges
-	}
-	// Citadel config
-	if remoteconfig.Spec.Citadel.Image == "" {
-		remoteconfig.Spec.Citadel.Image = defaultCitadelImage
-	}
-	if remoteconfig.Spec.Citadel.ReplicaCount == 0 {
-		remoteconfig.Spec.Citadel.ReplicaCount = defaultReplicaCount
-	}
-	// SidecarInjector config
-	if remoteconfig.Spec.SidecarInjector.Image == "" {
-		remoteconfig.Spec.SidecarInjector.Image = defaultSidecarInjectorImage
-	}
-	if remoteconfig.Spec.SidecarInjector.ReplicaCount == 0 {
-		remoteconfig.Spec.SidecarInjector.ReplicaCount = defaultReplicaCount
-	}
-}
